ref database comparison notebook

+++ tax_credit/process_mocks.py
@@ -10,8 +10,9 @@
 
 from os import path, makedirs
 from os.path import exists, join, basename
-from shutil import copyfile
+from shutil import copyfile, rmtree
 from urllib.request import urlretrieve
+from urllib.error import URLError
 from skbio import TreeNode, io
 import biom
 from biom import load_table
@@ -102,8 +103,11 @@ def extract_mockrobiota_data(communities, community_metadata, reference_dbs,
         for file_url_dest in [(forward_read_url, 'sequences.fastq.gz'),
                               (index_read_url, 'barcodes.fastq.gz')]:
             destination = join(seqs_dir, file_url_dest[1])
-            if not exists(destination):
-                urlretrieve(file_url_dest[0], destination)
+            if not exists(destination) and file_url_dest[0] != 'NA':
+                try:
+                    urlretrieve(file_url_dest[0], destination)
+                except ValueError:
+                    print('Error retrieving {0}'.format(file_url_dest[0]))
 
         # new directory containing expected taxonomy assignments at each level
         expected_taxa_dir = join(expected_data_dir, community,
@@ -172,16 +176,20 @@ def batch_demux(communities,
         sample_md = join(community_dir, metadata_fn)
 
         # demultiplex
-        demux_to_plot_quality(seqs_dir = seqs_dir,
-                              sample_md = sample_md,
-                              community_dir = community_dir,
-                              index_col = index_col,
-                              rc_barcodes = demux_params[community][0],
-                              rc_map_barcodes = demux_params[community][1],
-                              demux_fn = demux_fn,
-                              summary_fn = summary_fn,
-                              qual_plot_fn = qual_plot_fn,
-                              n_qual_plots = n_qual_plots)
+        if demux_params[community][0] is True:
+            demux_to_plot_quality(seqs_dir = seqs_dir,
+                                  sample_md = sample_md,
+                                  community_dir = community_dir,
+                                  index_col = index_col,
+                                  rc_barcodes = demux_params[community][1],
+                                  rc_map_barcodes = demux_params[community][2],
+                                  demux_fn = demux_fn,
+                                  summary_fn = summary_fn,
+                                  qual_plot_fn = qual_plot_fn,
+                                  n_qual_plots = n_qual_plots)
+
+        else:
+            load_demux_seqs(community_dir, seqs_dir, demux_fn, sample_md)
 
         print("{0} complete".format(community))
 
@@ -196,6 +204,7 @@ def demux_to_plot_quality(seqs_dir,
                           summary_fn='demux_summary.qzv',
                           qual_plot_fn='demux_plot_qual.qzv',
                           n_qual_plots=1):
+
     '''raw fastq -> SampleData[SequencesWithQuality]
 
     Demultiplex raw fastq files, summarize demux, and plot qual scores.
@@ -233,6 +242,15 @@ def demux_to_plot_quality(seqs_dir,
                                    rev_comp_mapping_barcodes = rc_map_barcodes)
     demux_seqs.per_sample_sequences.save(join(community_dir, demux_fn))
 
+    visualize_qual(demux_seqs, community_dir, summary_fn,
+                   qual_plot_fn, n_qual_plots)
+
+
+def visualize_qual(demux_seqs, community_dir, summary_fn,
+                   qual_plot_fn, n_qual_plots=1):
+
+    '''visualize demux summary and fastq quality plots.'''
+
     # demultiplexing summary
     try:
         dsum = demux.visualizers.summarize(demux_seqs.per_sample_sequences)
@@ -247,6 +265,39 @@ def demux_to_plot_quality(seqs_dir,
     qualplot.visualization.save(join(community_dir, qual_plot_fn))
 
 
+def load_demux_seqs(community_dir, seqs_dir, demux_fn, sample_md):
+    '''Load demultiplexed sequences into artifact and write artifact.'''
+
+    # extract sample name
+    with open(sample_md, 'r') as md_in:
+        md_in.readline()
+        sample_id = md_in.readline().strip().split('\t')[0]
+
+    # assemble artifact components
+    tmpdir = join(seqs_dir, 'tmp')
+    seq_fn = '{0}_1_L001_R1_001.fastq.gz'.format(sample_id)
+    if not exists(tmpdir):
+        makedirs(tmpdir)
+    copyfile(join(seqs_dir, 'sequences.fastq.gz'),
+             join(tmpdir, seq_fn))
+    with open(join(tmpdir, 'MANIFEST'), 'w') as manifest:
+        manifest.write('sample-id,filename,direction\n')
+        manifest.write('{0},{1},forward'.format(sample_id, seq_fn))
+    with open(join(tmpdir, 'metadata.yml'), 'w') as yml:
+        yml.write('\n')
+
+    seqs = qiime.Artifact.import_data(\
+        "SampleData[SequencesWithQuality]", tmpdir)
+    rmtree(tmpdir)
+
+    seqs.save(join(community_dir, demux_fn))
+
+    # Visualize currently fails on artifacts containing single sample
+    # (seqs has no attribute 'per_sample_sequences')
+    # visualize_qual(seqs, community_dir, summary_fn,
+    #               qual_plot_fn, n_qual_plots)
+
+
 def denoise_to_phylogeny(communities,
                          mock_data_dir,
                          trim_params,
@@ -254,6 +305,7 @@ def denoise_to_phylogeny(communities,
                          rep_seqs_fn='rep_seqs.qza',
                          feature_table_fn='feature_table.qza',
                          summary_fn='feature_table_summary.qzv'):
+
     '''SampleData[SequencesWithQuality] -> FeatureData[Sequence] +
                                            FeatureTable[Frequency]
         denoise fastqs with dada2, create feature table, rep_seqs,
